nyan/lexer: reorganize Lexer internals
